Add deferred authorizer
This type caches an authorization request.  This makes it possible for
the core library to try authorization with the deferred authorizer
mutiple times, which will fail until the request has asynchronously
resolved.  This failure can be caught and handled by the actix layer.
An actual request is only be sent exactly once with this mechanism.