Implement resource endpoint
The implementation creates an owning copy of the authorization header
from the request, even if the request returns a borrowed copy. This
avoid some lifetime entanglements that would otherwise occur: Scope
determination expects a mutable request as well, and the backend
endpoint trait expects an internally stored reference to the request.