Rework rouille into wrapper types
The crate no longer declares the trait, nor does it declare the request
or response types. Hence, we need to make new-type wrappers instead. To
improve ergonomics there should likely be a trait to wrap requests with
methods, which could also precalculate and thus optimize the internal
representation if desired (measure this).