Turn iterations into NonZeroU32 as a compat hack
The interface of pbkdf::verify change in ring `0.14` from

> verify(.., u32, ..) // 0.13
> verify(.., NonZeroU32, ..) // 0.14

Luckily for us, `NonZeroU32` implements both the trivial
`Into<NonZeroU32>` as well as `Into<u32>`. By switching to an internal
representation of `NonZeroU32` we can then use an `into()` call to let
the compiler deduce the correct type on both versions of ring.