Replace ring::Unspecified with RegistrarError
There is one instance where Unspecified was replaced with a
PrimitiveError. On checking a password the hash is checked for internal
consistency and a failure to validate indicates that the policy was
changed during operation. This is however not the failure of the user
but should be an internal server error.