Begin rework of code grant to statemachine
Native state machines with input and output will take unfortunately
likely take quite a while to be introduced to Rust, if at all. Writing
everything with async is neither flexible enough to support arbitrary
input parameters nor does it fit as it infects every interface using it
or requires a choice of local executor. Generators will also not be
enough as we demand additional input but this input depends on some
output so it can not be precalculated (or requires a side channel to be
inject which is even more awkward).

So, write these as a true Mealy automaton. We begin with the Refresh
flow as it is one of the simplest but still uses multiple primitives.