Adjust sub-crate version
The justification is as such. For synchronous server crates the
interface of oxide-auth is not expected to change by much and the
preview can be treated as a release candidate for a first version of the
crates.

In contrast, for actix we are already aware that the bump to 2.0 is
right around the corner and that it will feature async support. So this
version, supporting 1.0, is NOT a release candidate or a preview. At the
same time there should be at least one useful released version. This is
maybe best described by a version below 0.1.