New flag to omit version check for tagging
The release procedure is to merge the release-branch PR and then publish
the crates, and finally push the tags to ensure that no tag points to a
commit which did not have the versions published. However, this means
that the tag is conceptually created after publishing the crates. So we
need to skip the unpublished version check.