Sketch an implementation for resource
In a major difference compared to the other state machine, it does not
borrow the request for the duration of the state machine. Instead, it
expects it in the input from time-to-time. This makes it possible for
the executor (and potentially the state machine) to inspect the request
mutably to determine the scopes (and the authenticate header). This is
only avoided at the moment by explicitely cloning this state in the
library executor. This is overhead I would like to avoid at some point
in the future.

This will also require some changes to the structure of the state
machine as the output needs to borrow from the input. However, I do not
expect major complications. Only state that needs to be kept over
multiple iterations of the state machine must be cloned.