Finish resource implementation
The key insight was to make its output only depend on the current state
and not be too shy to clone some data from the input. Then the rest of
the machine transition was completed very quickly. I am considering also
applying similar modifications to the other state machines as a
simplification.

Note that it no longer borrows the request for the duration of the
state. There were previous concerns that not binding the request in this
manner might result in misuse. However, the truly only interface for
which misuse resistance is the main drivers is the high-level trait
interface and not the implementation details. For example, we already
allow 'recovering' grants that are not related to the token in the
request. In fact, one might consider the borrowing interface more prone
to misuse as it requires workarounds and wrappers to keep the borrow
alive for the duration of the state machine while holding onto the
request.