Fix actix example 'losing' client state
Actix really makes use of multi-threading now. This in particular means
the client state is create on each thread and it is not predictable
which handles the request. The closure creating this state has hidden a
bug that caused the state to not be shared across those threads. Mark
the actual state `Clone` and use the same across all handlers to fix
this issue.