fix: Allow a client to specify the client_id in body when using Authenticate
Currently when a client uses Basic Auth but still specifies client_id as
a body parameter, this library will raise an invalid_request error.
However specifying client_id is completely acceptable to the standard,
even when using Basic Auth. In fact, some client libraries, like for
example the elixir oauth2 library always set the client_id parameter.

This Commit makes it so we ignore any credentials specified in the body,
if Basic Auth is used